+++ src/components/SideBar/SideBarElements.js
@@ -1,8 +1,8 @@
 import styled from "styled-components";
-import { Colors } from "../../../Theme";
+import { Colors } from "../../Theme";
 
 export const SideBarContainer = styled.div`
   background-color: ${Colors.black};
-  width: 5rem;
-  height: 100%;
+  width: 1rem;
+  height: 8rem;
 `;
