+++ src/index.css
@@ -1,5 +1,5 @@
 * {
-  font-family: 'Comfortaa';
+  font-family: "Comfortaa";
 }
 
 body {
