--- src/pages/cashier/OrderPage/OrderPage.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-function OrderPage() {
-  return <div>OrderPage</div>;
-}
-
-export default OrderPage;
