--- src/pages/cashier/OrderPage/OrderPageElements.js
@@ -1,15 +1,0 @@
-import styled from "styled-components";
-import { Colors } from "../../Theme";
-
-export const Main = styled.div`
-  background-color: ${Colors.white};
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-`;
-
-export const ContentContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  height: 100%;
-`;
